% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breasy_demo_data.R
\docType{data}
\name{breasy_demo_data}
\alias{breasy_demo_data}
\title{breasy_demo_data Example data set with simulated benefit-risk information}
\format{
A data frame with the following variables (in this order):
\describe{
  \item{ESTIMATE}{estimate}
  \item{ANALYSIS_SET}{analysis set}
  \item{OUTCOME}{outcome}
  \item{DATA_SCOPE}{data scope}
  \item{STUDY}{study or trial number}
  \item{SUBGROUP}{subgroup}
  \item{SUBLEVEL}{sublevel of subgroup}
  \item{NUMBER_EVENTS_VERUM}{number events in verum group}
  \item{NUMBER_PATIENTS_VERUM}{number patients in verum group}
  \item{NUMBER_EVENTS_COMP}{number events in comparator group}
  \item{NUMBER_PATIENTS_COMP}{number patients in comparator group}
  \item{EFFECT_EXCESS}{effect excess}
  \item{LOWER95}{lower 95 percent confidence interval value of effect}
  \item{UPPER95}{upper 95 percent confidence interval value of effect}
  \item{NNT}{number needed}
  \item{DATE_EXCEL_GENERATED}{date excel generated}
  \item{STRATUM}{stratum variable}
}
}
\usage{
breasy_demo_data
}
\description{
breasy_demo_data Example data set with simulated benefit-risk information
}
\keyword{datasets}
