% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_breasy_forestplot.R
\name{breasy_forestplot}
\alias{breasy_forestplot}
\title{draw a forestplot grouped by Efficacy & Safety}
\usage{
breasy_forestplot(
  forest_data,
  excess_number = TRUE,
  incidence_values = TRUE,
  NNT = TRUE,
  title = NULL,
  lower_limit = NA,
  upper_limit = NA,
  efficacy_color = "#0091DF",
  safety_color = "#66B512",
  legend_color = "#D30F4B",
  sorting = "As Input",
  data_scope = NULL,
  ref_line_position = 0
)
}
\arguments{
\item{forest_data}{data set with required columns: EFFECT_xy, LOWERxy, UPPERxy, OUTCOME}

\item{excess_number}{logical value if a column of excess numbers should be drawn [default: TRUE]}

\item{incidence_values}{logical value if a column of incidence values should be drawn [default: TRUE]}

\item{NNT}{logical value if a column of NNT values should be drawn [default: TRUE]}

\item{title}{plot title [default: NULL]}

\item{lower_limit}{lower x-axis value [default: NULL]}

\item{upper_limit}{upper x-axis value [default: NULL]}

\item{efficacy_color}{color for the Efficacy font and Confidence Interval}

\item{safety_color}{color for the Safety font and Confidence Interval}

\item{legend_color}{color for the Legend font and Arrows}

\item{sorting}{Sorting variable (possible selections: "As Input", "Alphabetical", or "Effect") [default: "As Input"]}

\item{data_scope}{data scope variable}

\item{ref_line_position}{reference line position}
}
\description{
draw a forestplot grouped by Efficacy & Safety
}
