% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_effect_calc_new.R
\name{effect_calc_new}
\alias{effect_calc_new}
\title{file creation function for BReasy - Creates BReasy input from SAS files}
\usage{
effect_calc_new(
  data = data,
  effect = effect,
  day = "7",
  param = param,
  outcome = outcome,
  scope = scope,
  datascope = datascope,
  population = population,
  population_value = population_value,
  treatment = treatment,
  verum = verum,
  comparator = comparator,
  cnsr = cnsr,
  event = event,
  strat = strat,
  subgroup = subgroup,
  aval = aval
)
}
\arguments{
\item{data}{data frame with adtte data (merged with adsl data)}

\item{effect}{effect estimate (possible selection: CID, EXCESS_CID, ARD, EXCESS_ARD, HR, IRD, EXCESS_IRD)}

\item{day}{integer value for CID calculation}

\item{param}{character with parameter variable}

\item{outcome}{character(s) with outcome from param}

\item{scope}{character with scope variable}

\item{datascope}{character with scope selection}

\item{population}{character with population flag variable}

\item{population_value}{numeric or character with population flag value}

\item{treatment}{character with treatment variable}

\item{verum}{character(s) with treatment selection for verum}

\item{comparator}{character(s) with treatment selection for comparator}

\item{cnsr}{character with censoring variable}

\item{event}{character with event selection from cnsr}

\item{strat}{character(s) with stratification variable}

\item{subgroup}{character(s) with subgroup variable}

\item{aval}{character with aval variable}
}
\description{
file creation function for BReasy - Creates BReasy input from SAS files
}
